Move noopConsumer to DistributedFunctions
